requirements.txt
torch
torchvision
facenet-pytorch
faiss-cpu
opencv-python
numpy
Pillow